+++ CHANGELOG.md
@@ -13,6 +13,14 @@ BUMP:major - Increases the major version (x.0.0)
 -->
 
 <!-- BUMP:minor -->
+## v2.3.0 - 2024-04-24
+### Changed
+- feat: `Walking time duration` - Adjusts schedule accordingly to the time it takes to walk from your location to the station. 
+
+### Deprecated
+- Removed color highlighting
+
+<!-- BUMP:minor -->
 ## v2.2.0 - 2024-04-22
 ### Changed
 - feat: Set activity for specific days / hours